Ported to subclassing faketensor
