+++ torch/fx/experimental/fake_symbolic_tensor.py
@@ -1,0 +1,105 @@
+import torch
+import sympy
+from torch._C import _disabled_torch_function_impl
+from torch._meta_registrations import meta_funcs, register_meta
+
+def create_contiguous(shape):
+    if len(shape) == 0:
+        return []
+    strides = [1]
+    for dim in reversed(shape[:-1]):
+        strides.append(dim * strides[-1])
+    return list(reversed(strides))
+
+class FakeSymbolicTensor(torch.Tensor):
+    @staticmethod
+    def __new__(cls, sym_shape, sym_strides, dtype, layout, requires_grad, device):
+        # sym_strides doesn't work yet
+        # TODO: this is wrong in general
+        offset = 0
+        contiguous_strides = create_contiguous(sym_shape)
+        r = torch.Tensor._make_wrapper_subclass(
+            cls, sym_shape,
+            contiguous_strides, offset,
+            dtype=dtype, layout=layout, requires_grad=requires_grad,
+            device=device
+        )
+        return r
+
+    __torch_function__ = _disabled_torch_function_impl
+
+
+    def new_empty(self, shape):
+        return FakeSymbolicTensor(shape, None, self.dtype, self.layout, self.requires_grad, self.device)
+
+    @classmethod
+    def __torch_dispatch__(cls, func_overload, types, args=(), kwargs=None):
+        if func_overload in meta_funcs:
+            return meta_funcs[func_overload](*args, **kwargs)
+
+        if func_overload == torch.ops.aten.new_empty.default:
+            self = args[0]
+            shape = args[1]
+            return FakeSymbolicTensor(shape, self.stride(), self.dtype, self.layout, self.requires_grad, self.device)
+
+        raise RuntimeError(f"operator {func_overload} not supported")
+
+
+class PySymInt(object):
+    def __init__(self, expr, shape_env):
+        self.expr = expr
+        self.shape_env = shape_env
+
+    def wrap(self, num):
+        return PySymInt(sympy.Integer(num), self.shape_env)
+
+    def __str__(self):
+        return f"PySymInt({self.expr})"
+
+    def __int__(self):
+        # import pdb; pdb.set_trace()
+        return self.shape_env.evaluate_expr(self.expr)
+
+    def __bool__(self):
+        return bool(self.shape_env.evaluate_expr(self.expr))
+
+magic_methods = {
+    'add': lambda a, b: a + b,
+    'radd': lambda a, b: a + b,
+    'sub': lambda a, b: a - b,
+    'mul': lambda a, b: a * b,
+    'div': lambda a, b: a / b,
+    'mod': lambda a, b: a % b,
+    'eq': lambda a, b: sympy.Eq(a, b),
+    'gt': lambda a, b: sympy.Gt(a, b),
+    'lt': lambda a, b: sympy.Lt(a, b),
+}
+
+for method, func in magic_methods.items():
+    method_name = f'{method}'
+    def create_magic_impl(func):
+        def magic_impl(self, other):
+            if isinstance(other, PySymInt):
+                other = other.expr
+            return PySymInt(func(self.expr, other), self.shape_env)
+        return magic_impl
+
+    # this should be wrapped transparently into torch._C.SymbolicIntNode
+    setattr(PySymInt, method_name, create_magic_impl(func))
+
+class ShapeEnv(object):
+    def __init__(self):
+        self.guards = []
+        self.shape_env = {}
+
+    def create_symint(self, name, val):
+        sympy_expr = sympy.Symbol(name)
+        py_sym_int = PySymInt(sympy_expr, self)
+        cpp_sym_int = torch._C.SymbolicIntNode.new_symint(py_sym_int)
+        self.shape_env[sympy_expr] = val
+        return cpp_sym_int
+
+    def evaluate_expr(self, expr):
+        concrete_val = expr.subs(self.shape_env)
+        self.guards.append((expr, concrete_val))
+        return concrete_val
